Move funcs from Predicates to PredicatesElectra
